refactor: consolidate V3Player components to features/player/components
