+++ frontend/webui/src/features/player/components/V3Player.safari.test.tsx
@@ -2,10 +2,10 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { render } from '@testing-library/react';
 import V3Player from './V3Player';
-import Hls from '../features/player/lib/hlsRuntime';
+import Hls from '../lib/hlsRuntime';
 
 // Mock HLS.js
-vi.mock('../features/player/lib/hlsRuntime', () => {
+vi.mock('../lib/hlsRuntime', () => {
   const HlsMock = vi.fn().mockImplementation(function (this: any) {
     return {
       on: vi.fn(),
