+++ frontend/webui/src/features/player/components/V3Player.serviceRef.test.tsx
@@ -3,9 +3,9 @@ import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 
 import { describe, it, expect, vi, afterEach, beforeEach } from 'vitest';
 import V3Player from './V3Player';
-import type { V3PlayerProps } from '../types/v3-player';
+import type { V3PlayerProps } from '../../../types/v3-player';
 
-vi.mock('../features/player/lib/hlsRuntime', () => {
+vi.mock('../lib/hlsRuntime', () => {
   const HlsMock = vi.fn().mockImplementation(function (this: any) {
     return {
       on: vi.fn(),
